requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.103.1
filelock==3.12.4
fsspec==2023.9.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.1
pydantic==2.3.0
pydantic_core==2.6.3
python-decouple==3.8
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.22.0